+++ module/core/database/migration/migrations/sqlite/000006_falabella_sync_variations.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS falabella_sync_status_variation;